Updated server setup script.

+++ setup_server.sh
@@ -2,7 +2,13 @@
 # Run as root
 
 # Do this only once!
-# echo "/home/mpi_fractal/fractal_cluster *(rw,async,no_root_squash,no_subtree_check)" >> /etc/exports
+if [ -z "$1" ]; then
+    if [ "$1" == "first" ]; then
+        echo "/home/mpi_fractal/fractal_cluster *(rw,async,no_root_squash,no_subtree_check)" >> /etc/exports
+    fi
+    echo "Unknown command!"
+    exit -1
+fi
 
 # NFS server must be started and firewall disabled
 systemctl stop nftables
